new class Block and class Grid
